gugu

+++ admin/makanan_proses.php
@@ -12,8 +12,8 @@
             $makanan=$_REQUEST['makanan'];
             $deskripsi_makanan=$_REQUEST['deskripsi_makanan'];
             
-            $asal = $_FILES['gambar']['tmp_name'];
-            $simpan_gambar = "../gambar/".$_FILES['gambar']['name']; 
+            $asal = $_FILES['gambar_makanan']['tmp_name'];
+            $simpan_gambar = "../gambar/".$_FILES['gambar_makanan']['name']; 
             move_uploaded_file($asal, $simpan_gambar);
             
             $makanan_input = mysqli_query($connection, "INSERT INTO 
@@ -29,41 +29,42 @@
             
         break;
 
-        case 'edit':
-            $id_makanan = $_REQUEST['id_makanan'];
-            $makanan = $_REQUEST['makanan'];
-            $deskripsi_makanan = $_REQUEST['deskripsi_makanan'];
-            $centang = isset($_REQUEST['centang']) ? $_REQUEST['centang'] : '0';
-            $simpan_gambar = '';
-    
-            // Jika checkbox dicentang, proses upload gambar baru
-            if ($centang == '1' && isset($_FILES['gambar_makanan'])) {
-            $asal = $_FILES['gambar']['tmp_name'];
-                $nama_file = $_FILES['gambar']['name'];
-                $simpan_gambar = "../gambar/".$nama_file;
-                move_uploaded_file($asal, $simpan_gambar);
-            } else {
-                // Ambil path gambar lama dari database jika tidak upload baru
-                $result = mysqli_query($connection, "SELECT gambar_makanan FROM tb_makanan WHERE id_makanan = '$id_makanan'");
-                $row = mysqli_fetch_assoc($result);
-                $simpan_gambar = $row['gambar'];
-            }
-            
-            // Query UPDATE yang benar
-            $query = "UPDATE tb_makanan SET 
-                      makanan = '$makanan', 
-                      deskripsi_makanan = '$deskripsi_makanan', 
-                      gambar_makanan = '$simpan_gambar' 
-                      WHERE id_makanan = '$id_makanan'";
-            
-            $makanan_edit = mysqli_query($connection, $query);
-            
-            if ($makanan_edit) {
-                echo "<script>alert('Update berhasil')</script>";
-            } else {
-                echo "<script>alert('Gagal update: ".mysqli_error($connection)."')</script>";
-            }
-        break;
+case 'edit':
+    $id_makanan = $_REQUEST['id_makanan'];
+    $makanan = $_REQUEST['makanan'];
+    $deskripsi_makanan = $_REQUEST['deskripsi_makanan'];
+    $simpan_gambar = "";
+
+    $centang = $_REQUEST['centang'] ?? '0';
+    $gambar_column_exists = false;
+    $result = mysqli_query($connection, "SHOW COLUMNS FROM tb_makanan LIKE 'gambar_makanan'");
+    if ($result && mysqli_num_rows($result) > 0) {
+        $gambar_column_exists = true;
+    }
+
+    if ($centang == '1' && isset($_FILES['gambar_makanan']) && $_FILES['gambar_makanan']['error'] == 0 && $gambar_column_exists)
+    {
+        $asal = $_FILES['gambar_makanan']['tmp_name'];
+        $simpan_gambar = "../gambar/" . $_FILES['gambar_makanan']['name'];
+        move_uploaded_file($asal, $simpan_gambar);
+
+        $makanan_edit = mysqli_query($connection, "UPDATE tb_makanan SET id_makanan = '$id_makanan', makanan = '$makanan', deskripsi_makanan = '$deskripsi_makanan', gambar_makanan = '$simpan_gambar' WHERE id_makanan = '$id_makanan'");
+    }
+    else
+    {
+        $makanan_edit = mysqli_query($connection, "UPDATE tb_makanan SET id_makanan = '$id_makanan', makanan = '$makanan', deskripsi_makanan = '$deskripsi_makanan' WHERE id_makanan = '$id_makanan'");
+    }
+
+    switch ($makanan_edit) {
+        case true:
+            echo "<script>alert ('update in db')</script>";
+            break;
+        default:
+            echo "<script>alert ('input in db failed')</script>";
+            break;
+    }
+
+    break;
 
         case 'hapus':
             $id_makanan = $_REQUEST['id_makanan'];
